IbWriteInterestQueue: Bugfix
This one caused the queue to hit it's limit which should not
be possible. Actually, the queue wasn't full but the
resBack ptr was incremeneted multiple times and was in front of
the back pointer. This might happen after a few seconds on the
first benchmark or after multiple benchmarks.
I haven't found the actual problem on the old code, so I rewrote
it and it works fine now.