ib: Refactor connection creation
The native buffers for the ORB are now allocated when a new
connection is created and not created on init to save memory.
Thus, we need to actively force connection creation once dxnet
requires a connection to be created. This requires some changes
to the JNI bindings as well as the logic calling to the native
subsystem